requirements.txt
torch>=2.1
numpy>=1.23
faiss-cpu>=1.7.4; platform_system != "Windows"
